+++ Sources/SkillsMaster/Utilities/Constants.swift
@@ -23,6 +23,7 @@ enum Constants {
             case .trae:        Color(red: 0.20, green: 0.94, blue: 0.55)  // Bright Green #32F08C
             case .hermes:      Color(red: 0.94, green: 0.45, blue: 0.15)  // Hermès Orange #F07326
             case .workBuddy:   Color(red: 0.55, green: 0.89, blue: 0.77)  // Mint #8CE3C4
+            case .zcode:       Color(red: 0.16, green: 0.48, blue: 0.92)  // Azure #297AEB
             }
         }
     }